requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.5.0
appnope==0.1.2
asgiref==3.5.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
ccxt==1.77.41
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
ciso8601==2.2.0
click==8.0.3
cryptography==36.0.2
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
distlib==0.3.4
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
filelock==3.4.2
fonttools==4.31.1
frozenlist==1.3.0
gevent==21.12.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
ipykernel==6.9.2
ipython==8.1.1
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.1.2
jupyter-core==4.9.2
kiwisolver==1.4.0
matplotlib==3.5.1
matplotlib-inline==0.1.3
multidict==6.0.2
nest-asyncio==1.5.4
numpy==1.22.2
openpyxl==3.0.9
packaging==21.3
pandas==1.4.1
parso==0.8.3
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.4.1
plotly==5.6.0
printy==2.2.0
prompt-toolkit==3.0.28
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycares==4.1.2
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
stack-data==0.2.0
starlette==0.19.0
statsmodels==0.13.2
TA-Lib==0.4.24
tenacity==8.0.1
threadpoolctl==3.1.0
tornado==6.1
traitlets==5.1.1
urllib3==1.26.9
virtualenv==20.13.1
wcwidth==0.2.5
websocket==0.2.1
websocket-client==1.3.2
yarl==1.7.2
zope.event==4.5.0
zope.interface==5.4.0
